added email notification

+++ transaction/views.py
@@ -1,30 +1,34 @@
 from django.shortcuts import render, redirect
-from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth import authenticate
+from django.core.mail import send_mail
 import random
 from django.contrib import messages
-from datetime import datetime
 from django.core.paginator import Paginator
+from decouple import config
+from transaction.models import (
+    Deposit,
+    AccountDetail,
+    Withdraw,
+    Transfer,
 
-from transaction.models import Deposit, AccountDetail, Withdraw, Transfer
+)
+
 from transaction.forms import DepositForm, WithdrawForm, TransferForm
 from transaction import forms
 from . import models
 from account.models import Account
 
+
 # Create your views here.
-
-
 def randomNum():
     return int(random.uniform(1000000000, 9999999999))
-
 
 def account_view(request):
     user = request.user
     if user.is_authenticated:
         try:
-            active_user = AccountDetail.objects.get(email=user)
+            active_user = AccountDetail.objects.get(email=request.user)
         except:
-            my_field = Account._meta.get_field('email')
             active_user = AccountDetail()
             active_user.account_number = randomNum()
             active_user.account_type = 'savings'
@@ -44,21 +48,33 @@ def deposit_view(request):
             form = forms.DepositForm(request.POST)
             if form.is_valid():
                 form.save()
-                actii_user = models.AccountDetail.objects.get(
-                    email=request.user)
-                acti_user = models.Deposit.objects.filter(
-                    email=request.user).last()
-                depositor_account_number = acti_user.dep_account
-                temp = acti_user
 
-                # act_user = models.Status.objects.get(account_number = depositor_account_number)
-                act_user_amout = acti_user.amount
-                actii_user.balance = actii_user.balance + act_user_amout
+                actii_user = models.AccountDetail.objects.get(email=request.user)
+                acti_user = models.Deposit.objects.filter(dep_email=request.user).last()
 
-                actii_user.save()
-                # temp.delete()
-                messages.success(request, "deposited successfully")
-                return redirect('deposit-history')
+                if (actii_user.account_number != acti_user.dep_account):
+                    messages.error(request, "Your account number is not correct")
+                    return redirect('deposit')
+                else:
+                    act_user_amout = acti_user.amount
+                    actii_user.balance = actii_user.balance + act_user_amout
+
+                    # send mail
+                    mail_name = 'Django Bank'
+                    sender = 'isaiaholadapo18@gmail.com'
+                    message = '₦' + str(act_user_amout) + " deposited, your current balance is ₦" + str(actii_user.balance)
+                    send_mail(
+                        'Contact Form mail from ' + mail_name,
+                        'Sender email: ' + ' ' + sender + '\n' + message,
+                        sender,
+                        [acti_user.dep_email, ]
+                    )
+
+                    actii_user.save()
+                    messages.success(request, "deposited successfully")
+
+                    return redirect('deposit-history')
+
         else:
             form = forms.DepositForm()
     else:
@@ -81,26 +97,41 @@ def withdraw_view(request):
                 temp = withdrawal_user
                 withdrawal_user = models.AccountDetail.objects.get(
                     email=request.user)
-                wbal = withdrawal_user.balance
-                if withdrawal_amount < 5000:
-                    temp.delete()
-                    messages.error(
-                        request, "You can't withdraw amount less than 5,000")
+                if withdrawal_account != withdrawal_user.account_number :
+                    messages.error(request, "Your account number is not correct")
                     return redirect('withdraw')
-                elif withdrawal_amount > 50000:
-                    temp.delete()
-                    messages.error(
-                        request, "You can't withdraw amount more than 50,000")
-                    return redirect('withdraw')
-                elif wbal > withdrawal_amount:
-                    withdrawal_user.balance = withdrawal_user.balance - withdrawal_amount
-                    withdrawal_user.save()
-                    messages.success(request, "Withdrawal successful")
-                    return redirect('withdraw-history')
                 else:
-                    temp.delete()
-                    messages.error(request, "You don't have enough Money.")
-                    return redirect('withdraw')
+                    wbal = withdrawal_user.balance
+                    if withdrawal_amount < 50:
+                        temp.delete()
+                        messages.error(
+                            request, "You can't withdraw amount less than 50")
+                        return redirect('withdraw')
+                    elif withdrawal_amount > 5000000:
+                        temp.delete()
+                        messages.error(
+                            request, "You can't withdraw amount more than 5,000,000")
+                        return redirect('withdraw')
+                    elif wbal > withdrawal_amount:
+                        withdrawal_user.balance = withdrawal_user.balance - withdrawal_amount
+                        # send mail
+                        mail_name = 'Django Bank'
+                        sender = 'isaiaholadapo18@gmail.com'
+                        message = '₦' + str(withdrawal_amount) + " withdrawn, your current balance is ₦" + str(
+                            withdrawal_user.balance)
+                        send_mail(
+                            'Contact Form mail from ' + mail_name,
+                            'Sender email: ' + ' ' + sender + '\n' + message,
+                            sender,
+                            [withdrawal_user.email, ]
+                        )
+                        withdrawal_user.save()
+                        messages.success(request, "Withdrawal successful")
+                        return redirect('withdraw-history')
+                    else:
+                        temp.delete()
+                        messages.error(request, "You don't have enough Money.")
+                        return redirect('withdraw')
             else:
                 form = forms.WithdrawForm()
     else:
@@ -126,19 +157,42 @@ def transfer_view(request):
                     account_number=receiver_account_no)
                 sender = models.AccountDetail.objects.get(
                     email=request.user)
-                if sender.balance > transfer_amount:
-                    sender.balance = sender.balance - transfer_amount
-                    receiver_account.balance = receiver_account.balance + transfer_amount
-
-                    sender.save()
-                    receiver_account.save()
-                    messages.success(request, "Transaction successful")
-                    return redirect('transfer-history')
-                else:
-                    temp.delete()
-
-                    messages.error(request, "You don't have enough Money.")
+                if receiver_account == sender.account_number:
+                    messages.error(
+                        request, "You can't transfer money to yourself")
                     return redirect('transfer')
+                else:
+                    if sender.balance > transfer_amount:
+                        sender.balance = sender.balance - transfer_amount
+                        receiver_account.balance = receiver_account.balance + transfer_amount
+                        # sender mail
+                        mail_name = 'Django Bank'
+                        senderm = 'isaiaholadapo18@gmail.com'
+                        message = '₦' + str(transfer_amount) + " withdrawn, your current balance is ₦" + str(sender.balance)
+                        send_mail(
+                            'Contact Form mail from ' + mail_name,
+                            'Sender email: ' + ' ' + senderm + '\n' + message,
+                            senderm,
+                            [sender.email, ]
+                        )
+                        # receiver mail
+                        message = '₦' + str(transfer_amount) + " received, your current balance is ₦" + str(
+                            sender.balance)
+                        send_mail(
+                            'Contact Form mail from ' + mail_name,
+                            'Sender email: ' + ' ' + senderm + '\n' + message,
+                            senderm,
+                            [receiver_account.email, ]
+                        )
+                        sender.save()
+                        receiver_account.save()
+                        messages.success(request, "Transaction successful")
+                        return redirect('transfer-history')
+                    else:
+                        temp.delete()
+
+                        messages.error(request, "You don't have enough Money.")
+                        return redirect('transfer')
         else:
             form = forms.TransferForm()
     else:
@@ -151,16 +205,26 @@ def transaction_view(request):
     if not user.is_authenticated:
         return redirect('login')
     deposit = models.Deposit.objects.all().order_by('date').reverse()
-    deposit_paginator = Paginator(deposit, 5)
-    page_num = request.GET.get('page')
-    page = deposit_paginator.get_page(page_num)
+    transfer = models.Transfer.objects.all().order_by('date').reverse()
+    #deposit_paginator = Paginator((deposit, transfer), 5)
+    #page_num = request.GET.get('page')
+   # page = deposit_paginator.get_page(page_num)
     context = {
-        "depositi": deposit,
-        "page": page,
-        "count": deposit_paginator.count,
+       "depositi": deposit,
+        "transfer": transfer,
+        #"page": page,
+        #"count": deposit_paginator.count,
 
     }
     return render(request, 'transaction/transaction.html', context=context)
+
+
+def history_view(request):
+    history = models.History.objects.all().order_by('date').reverse()
+    context = {
+        "history": history
+    }
+    return render(request, 'transaction/history.html', context)
 
 
 def withdraw_history_view(request):
@@ -170,30 +234,32 @@ def withdraw_history_view(request):
     if request.method == "POST":
         fromdate = request.POST.get('fromdate')
         todate = request.POST.get('todate')
-        searchresult = Withdraw.objects.raw('select id,withdraw_amount,date from transaction_withdraw where date between "'+fromdate+'" and "'+todate+'"')
+        searchresult = Withdraw.objects.raw(
+            'select id,withdraw_amount,date from transaction_withdraw where date between "' + fromdate + '" and "' + todate + '"')
         searchresult_paginator = Paginator(searchresult, 5)
         page_num = request.GET.get('page')
         page = searchresult_paginator.get_page(page_num)
-        
+
         contexti = {
             'searchresult': searchresult,
-            'page' : page,
+            'page': page,
             'count': searchresult_paginator.count,
         }
-        
+
         return render(request, 'transaction/withdraw_history.html', contexti)
     else:
         withdraw = Withdraw.objects.all().order_by('date').reverse()
         withdraw_paginator = Paginator(withdraw, 5)
         page_num = request.GET.get('page')
         page = withdraw_paginator.get_page(page_num)
-        
+
         contexti = {
             'withdrawal': withdraw,
-            'page' : page,
+            'page': page,
             'count': withdraw_paginator.count,
         }
         return render(request, 'transaction/withdraw_history.html', contexti)
+
 
 def deposit_history_view(request):
     user = request.user
@@ -202,12 +268,13 @@ def deposit_history_view(request):
     if request.method == "POST":
         fromdate = request.POST.get('fromdate')
         todate = request.POST.get('todate')
-        searchresult = Deposit.objects.raw('select id,amount,date from transaction_deposit where date between "'+fromdate+'" and "'+todate+'"')
+        searchresult = Deposit.objects.raw(
+            'select id,amount,date from transaction_deposit where date between "' + fromdate + '" and "' + todate + '"')
         searchresult_paginator = Paginator(searchresult, 5)
         page_num = request.GET.get('page')
         page = searchresult_paginator.get_page(page_num)
         context = {
-            "searchresult" : searchresult,
+            "searchresult": searchresult,
             "page": page,
             "count": searchresult_paginator.count,
         }
@@ -218,26 +285,28 @@ def deposit_history_view(request):
         page_num = request.GET.get('page')
         page = deposit_paginator.get_page(page_num)
         context = {
-            "deposit" : deposit,
+            "deposit": deposit,
             "page": page,
             "count": deposit_paginator.count,
         }
         return render(request, 'transaction/deposit_history.html', context)
 
+
 def transfer_history_view(request):
-    user =  request.user
+    user = request.user
     if not user.is_authenticated:
         return redirect('login')
     if request.method == "POST":
         fromdate = request.POST.get("fromdate")
         todate = request.POST.get("todate")
-        searchresult = Transfer.objects.raw('select id,amount,date from transaction_transfer where date between "'+fromdate+'" and "'+todate+'"')
+        searchresult = Transfer.objects.raw(
+            'select id,amount,date from transaction_transfer where date between "' + fromdate + '" and "' + todate + '"')
         transfer_paginator = Paginator(searchresult, 5)
         page_num = request.GET.get('page')
         page = transfer_paginator.get_page(page_num)
         context = {
             "searchresult": searchresult,
-            "page" : page,
+            "page": page,
             "count": transfer_paginator.count,
         }
         return render(request, 'transaction/transfer_history.html', context)
@@ -248,30 +317,31 @@ def transfer_history_view(request):
         page = transfer_paginator.get_page(page_num)
         context = {
             "transfer": transfer,
-            "page" : page,
+            "page": page,
             "count": transfer.count,
         }
 
         return render(request, 'transaction/transfer_history.html', context)
 
 
-
 def deposit_search_view(request):
     if request.method == "POST":
         fromdate = request.POST.get('fromdate')
         todate = request.POST.get('todate')
-        searchresult = Deposit.objects.raw('select id,amount,date from transaction_deposit where date between "'+fromdate+'" and "'+todate+'"')
+        searchresult = Deposit.objects.raw(
+            'select id,amount,date from transaction_deposit where date between "' + fromdate + '" and "' + todate + '"')
         return render(request, 'transaction/searchdate.html', {"ssdata": searchresult})
     else:
         deposit_date = Deposit.objects.all()
         return render(request, 'transaction/searchdate.html', {"ssdata": deposit_date})
 
+
 def interest_view(self):
     user = request.user
     if not user.is_authenticated:
         return redirect('login')
-    int_user = models.AccountDetail.objects.get(email = request.user)
-    int_user.balance = int_user.balance * (1/ 0.01)
+    int_user = models.AccountDetail.objects.get(email=request.user)
+    int_user.balance = int_user.balance * (1 / 0.01)
     int_user.save()
     return render(request, 'transaction/home.html', {"int_user": int_user})
-    
+
